KAN-32 rename: 작업공간 페이지와 작품플래너 페이지를 route group로 그룹화
