+++ src/app/(after-login)/planner/[id]/page.module.scss
