+++ src/app/(after-login)/workspace/[id]/_components/index-pannel/IndexPannel.module.scss
