+++ src/app/(after-login)/workspace/[id]/page.module.scss
