+++ src/app/(after-login)/workspace/[id]/page.tsx
@@ -2,14 +2,14 @@
 
 import { useRef } from 'react'
 
-import { Editor } from '@tiptap/react'
+import { HandleEditor } from 'types/common/editor'
 
 import DefaultEditor from '@components/editor/DefaultEditor'
 
-import IndexPannel from './components/index-pannel/IndexPannel'
-import MemoPannel from './components/memo-pannel/MemoPannel'
-import PlannerPannel from './components/planner-pannel/PlannerPannel'
-import WorkspaceActionBar from './components/workspace-action-bar/WorkspaceActionBar'
+import IndexPannel from './_components/index-pannel/IndexPannel'
+import MemoPannel from './_components/memo-pannel/MemoPannel'
+import PlannerPannel from './_components/planner-pannel/PlannerPannel'
+import WorkspaceActionBar from './_components/workspace-action-bar/WorkspaceActionBar'
 
 import classNames from 'classnames/bind'
 
@@ -21,12 +21,8 @@ const cx = classNames.bind(styles)
  * MEMO(Sohyun) 목차 패널, 메모 패널은 에디터와 밀접하게 연결되어 있으므로 하위 컴포넌트로 묶어서 분리할 필요성 고민
  */
 
-export interface RefEditor {
-  getEditor: () => Editor | null
-}
-
 export default function WorkSpacePage() {
-  const editorRef = useRef<RefEditor>(null)
+  const editorRef = useRef<HandleEditor>(null)
 
   const handleSave = async () => {
     if (editorRef.current) {
